move table creation to activation hoook remove extra files remove phpcs ignore statements
